[REF] composer: extract selection from content editable helper
Backport of #5494

The goal is to re-use those functions in the test suite to easily
insert new text at the current selection (see next commit)

Task: 4592612
